# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GBF ERP API
  version: "0.2.0"
  description: |
    通用贸易 MVP 后端 HTTP 接口（**精简说明**）。完整行为以 `src/routes` 与 Zod schema 为准。

    - 成功响应多为 `application/json`；错误体由全局错误处理统一包装。
    - 鉴权：除 `/health`、`/api/auth/login`、`/api/auth/refresh`（及开发态 `x-username`）外，一般需 `Authorization: Bearer <access_token>`。
servers:
  - url: http://localhost:3100
    description: 本地默认端口
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: 登录（返回 access + refresh）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: token 等
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/auth/me:
    get:
      tags: [auth]
      summary: 当前用户
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/auth/permissions:
    get:
      tags: [auth]
      summary: 当前用户权限列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 权限字符串数组
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: 刷新 access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: 新 token
  /api/auth/logout:
    post:
      tags: [auth]
      summary: 撤销 refresh 会话
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: OK
...
  /api/approvals/pending:
    get:
      tags: [approvals]
      summary: 待我审批（需对应 approve 权限）
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 混合类型列表
  /api/approvals/{orderType}/{id}/timeline:
    get:
      tags: [approvals]
      summary: 审批时间线
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderType
          in: path
          required: true
          schema: { type: string, enum: [purchase, sales] }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 审计时间线
  /api/config/webhooks:
    get:
      tags: [config]
      summary: Webhook 端点列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [config]
      summary: 创建 Webhook
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Created
  /api/config/webhooks/{id}:
    delete:
      tags: [config]
      summary: 删除 Webhook
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access token（`typ=access`）
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
    LoginBody:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3 }
        password: { type: string, minLength: 6 }
...